Compute sequence of 2 events – OMOP

Compute sequence of 2 events – OMOP

Usage

compute_event_sequence_omop(
  cohort,
  grouped_list,
  site_col,
  user_cutoff = 30,
  n_event_a = 1,
  n_event_b = 1,
  patient_level_tbl = FALSE,
  time = FALSE,
  time_period = "year",
  time_span = c("2011-01-01", "2023-12-31"),
  event_csv = patienteventsequencing::pes_event_file
)

Arguments

cohort

a table with the cohort of interest; should minimally include:

site

person_id

start_date

end_date

grouped_list

list of columns in the data that should be used to group the results

site_col

the name of the column with site name(s)

user_cutoff

a user specified threshold against which to measure the number of days between the two events

n_event_a

the number of times Event A should occur before establishing the index date

n_event_b

the number of times Event B should occur before esrablishing the occurrence date

patient_level_tbl

a logical to indicate whether an intermediate table with patient level results should be output. if TRUE, a dataframe with one row per patient will be returned in addition to aggregate output

time

a logical that tells the function whether you would like to look at the output over time

time_period

when time = TRUE, this argument defines the distance between dates within the specified time period. defaults to year, but other time periods such as month or week are also acceptable

time_span

when time = TRUE, this argument defines the start and end dates for the time period of interest. should be formatted as c(start date, end date) in yyyy-mm-dd date format.

event_csv

a csv file with the definitions for each event with the columns

event - A or B

event_label - a descriptive label for the event

domain_tbl - the default CDM table from which data is retrieved

concept_field - the field in the table where the codes of interest are stored

date_field - the date field to be used to establish the index & occurrence dates

codeset_name - the name of the codeset in the specs directory to define the variable of interest

filter_logic - a string indicating any filter logic that should be applied to establish the event ex: an Hba1c > 6.5

Value

an aggregated dataframe outlining the number of patients that had the two events occur in X number of days

    if patient_level_tbl = TRUE, a patient level dataframe is also returned